Move Session into a separate file.
